pin to a specific nightly
Given that we use a lot of nightly feature, we want to secure against
breaking changes by pinning a specific version. That way we don't rely
on whatever nightly version the user has installed their system.
Additionally this ensures that every commits compiles even in the future
because the nightly version is explicitly set.