make syscalls ABI aware
Depending on the ABI used to invoke a syscall pointers point to values with
different sizes and layouts. A trivial example of this is `IoVec`: On Amd64 the
contained data pointer and length are 64-bit values, but on i386 these are
32-bit values.
To that end we seperate values from their in-memory representation and convert
back and forth at the time the memory operation is done. The syscall handler
are not involved in this process at all other than that they need to pass the
ABI as a parameter.

Note that the ABI does not always respond the loaded binary: 64-bit processes
can still use the i386 ABI through the `int 0x80` instruction.