add zero copy IO
This commit makes it possible for the read and write syscalls to
execute without making unneeded copies to a buffer first. This also
makes it possible for larger reads and writes to succeed because
they're no longer limited to the size of the buffer. This also greatly
improves the performance for file descriptors that return no data on
reads or ignore writes.