lots of fs changes
Nodes are no longer just Files, Directories and Symlinks, but any type
that implements INode. Along with that INode now allows for custom
implementations of open which is needed for things like fdfs.
Other than that I moved some files around so that fs stuff is placed
next to each other.