get rid of ActiveVirtualMemory
This type was mostly intended to support switching back and forth
between several virtual memories, but it turns out that we don't
really need this. Instead we lazily update Cr3 whenever we do an
operation in a given virtual memory. The key difference is that we
try to keep these operations short (e.g. map a page or read some bytes)
and don't encode in the type system that a virtual memory is active for
an extended time. Each operation takes a guard to make sure we don't
try to do work on two virtual memory at the same time. With that
implemented we can get rid of ActiveVirtualMemory and all its ugliness
(especially around async).