rework errors
The old constructors that took a tuple were always quite an eye sore.
The intent behind the tuple was to prevent to prevent eta reductions
that mess with the call site (The call site is recorded for debugging).
Instead of a function call, we now use the `err!` macro to construct
errors. The `bail!` and `ensure!` macros are used to shorten common
patterns even more and provide semantic meaning (both macros are
inspired by the anyhow crate).