declare SECRETS as extern static
Previously when we tried to declare SECRETS as a regular static, the
compiler realized that we never wrote the SECRETS and so it just
assumed that all reads would return the value we initialized it with.
In order to make sure that we're actually reading the value the SEV
firmware writes to the SECRETS page, we used volatile reads to make
sure the compiler didn't optimize the reads out.
Declaring the secrets page as a extern static achieves the same thing
and also makes it much clearer that this value comes isn't initialized
by Rust code.