handle #VC directly in the workload kernel
It turned out that only handling #VC in the supervisor was a
performance bottleneck. Previously the workload kernel communicated
with the supervisor by causing #VC exceptions. This has been replaced
by "supervisor services". The workload pushes commands into a command
buffer that's shared with the supervisor and the supervisor then
asynchronously handles these commands.