make it possible to supply multiple inputs
The key difference between having multiple smaller inputs vs one larger
input is that every smaller input is hashed separately as opposed to
one hash over the entire input. This makes it possible to easily append
to inputs without having to calculate the hash over the entire input.
This is very useful in scenarios where a verifier doesn't know an
input, but knows its hash. With these changes, it's possible for a
verifier to verify an attestation report with multiple such inputs.