initialize LSTAR only once
Previously we have to initialize it everytime because the compiler was
allowed to instantiate the inline assembly more than once. This is
technically still true, but doesn't matter much because there's no
surrounding relevant context (e.g. code in a function the asm block was
inlined into that could be present for normal inline assembling block).