implement multi-threaded supervisor for SNP
We already have a multi-threaded supervisor for TDX. Injecting
interrupts into the workload kernel is kind of a pain with a single-
threaded supervisor. If the targeted VMSA is already in use it, writing
to it will just block the supervisor. This makes high frequency timer
interrupts across many vCPUs pretty much impossible.