add support for handling timer irqs in the kernel
This is the first step for preemptive scheduling. Some processes will
never invoke any syscalls on their own and so the scheduler never has a
chance to run another thread. Similarly, we never have a chance to
check if a thread is still supposed to be running if the thread never
yields to the kernel.
To solve this, we now expect the supervisors to periodically inject
timer interrupts and force a transition out of userspace into
kernelspace.