implement id block signing
ID block signatures are required for attestation security. They're not
optional.
Starting with this commit, we require ID block and ID block signatures.
The public key is derived using public key recovery and a fixed
signature. This scheme allows us to have a signing key without a
trusted party that holds the private key.