add "Link" concept
A "Link" is an inode combined with a path to the inode. Hard-linked
files can be in multiple directories, so without this its not easy to
get a path to a file. The path is tracked in a LinkLocation struct.
When an inode is renamed/moved, the LinkLocation is automatically
updated with the new location. This allows tracking the path even after
it moved.
This is a prerequisite for inotify.