implement sparse buffers
Some workloads use very large files with holes in them. In those cases,
storing the pages in a vector isn't feasible. Instead, store them in a
BTreeMap. This is slower though, so we don't want to do that for all
buffers, but just the ones that need to be sparse.