fix race condition in timer code
Quite frankly, I don't remember why this suddenly fixes things. I know
that I know what was wrong when I wrote the fix, but I don't remember.
I realize that that's not great and that the timer code is admittedly
already a bit too complicated to be maintainable. Let's hope we don't
have to change it soon.