clean up after cancelled futex ops
Some program use a lot of futex ops that time out frequently.
Previously if that happened, we'd just keep appending to the list of
wait objects and only remove entries when the futex was woken up.
Instead, we now also remove entries when a futex op is cancelled. This
is implemented with a linked list.