hold weak reference in epoll interest list
The references in the epoll interest list need to be even weaker than
regular (non-strong) file descriptors because we must allow for the ofd
to be dropped while its still on an interest list. Some applications
depend on the write half of a pipe being closed while its still on an
epoll interest list.