fix partial sendfile ops
Previously we didn't handle errors very gracefully. For example, we
would sometimes return an error even though some bytes were transfered.
To make things easier with updating the file offset for the input fd,
we know assume that the input fd supports mmap and then update the file
offset explicitely. We are allowed to do this.