use a single slot for dynamic memory on SNP
It turns out that having many slots is very slow on SNP. Destroying
each slot takes around 50ms regardless of the size. It's very much
possible that this is just a kernel bug that might be fixed in the
future. For now it's easier to just switch to a single slot.