fix futex performance
It turns out that waiting for confirmation every time is *really* slow.
I saw 70ms roundtrips for a simple WAIT+WAKE combo.

The reasons we confirmed wake ops in the first place was cancellations:
If a WAIT operation timed out and got woken at the same time, there was
a race condition where the timeout error was returned and the wake
operation was dropped. This lead to deadlocks. Instead, we now
atomically cancel the futex and check if it was woken up before
returning a timeout error. Note that there's still a chance for wakups
to get dropped when the process is killed, but that's not as much of an
issue. If it becomes an issue, we could consider detecting this
situation and waking up another WAIT op.