optimize Notify
Notify::wait is often called in a loop. This change makes it possible
to reuse the same NotifyWait instance for multiple iterations. This is
a bit faster because we can skip the setup code whenever an instance is
reused. It's also a bit more convenient and it's now possible to call
until directly on the returned Future.

This change optimizes Notify further by moving the generation counter
outside the Mutex and replacing it with an AtomicU64. The generation
counter is still only written whenever the lock is taken, but this
approach allows for optimistic reads without taking a lock. In a lot
of cases, polling the NotifyWait doesn't require *any* writes (
including writes necessary for taking locks) which should help greatly
reduce contention. Writes are only necessary when updating the list of
wakers.