reshape epoll interface
Instead of two functions (one of which just signals whether the other
is available), use a single function that returns a Result.
One big advantage of the new interface is that the returned object can
contain some additional information.
We'll use this for signalfd to store the thread that registered the fd.