run each unit test in a seperate thread
Next up, we'll add a bunch of tests for mmap-related syscalls. This is
much easier if the process only has a single thread and we don't have
to worry about conflicting operations between threads (e.g. one thread
intentionally leaving a hole in virtual memory that could be
unintentionally taken up by another thread).
Rust uses a single process per test if the panic mode is set to abort.