simplify virtual memory map updates
Previously, all the methods that updated the memory map had to deal
with having to split mappings. This required a bunch of repeated and
arguably too complicated code.
Instead shatter_mapping_at can be used to make sure that there is no
mapping containing a page that would need to be split. Once the mapping
has been shattered, the functions can simplify use any of the range
iterator methods provided by BTreeMap.
This will be come even more important once we implement mremap.