optimize stack mappings
Stack mappings are usually quite large. Large mappings are usually
handled by storing the pages in a BTreeMap to lessen to space wasted
by mapped-but-not-used pages. Stacks are special here because they have
special access pattern, starting from the top and growing down. We can
take advantage of this by storing the pages in a vector in reverse.
That way, the storage needed corresponds directly to how much of the
stack is in use.
Users usually don't map over the stack or change the protection
settings, so we don't need to implement fast-paths for splitting and
can simply revert back to dense mappings should we ever need to split.