fix creating files through symlinks
There are some nodes (like /proc/pid/fd/*) where reading the symlink
to follow it is not the right thing to do. Reading the symlink might
return a pseudo path (e.g. socket:[1]) instead of properly resolving
to the file.
That said, there are situations where the symlink cannot be resolved
(e.g. because the file does not exist) and so in that case it's
important to have the path.
create_file does different things depending on what succeeds.